pyproject.toml
[project]
name = "gaussx"
version = "0.0.9"
description = "Structured linear algebra, Gaussian distributions, and exponential family primitives for JAX."
authors = [
    {name="J. Emmanuel Johnson", email="jemanjohnson34@gmail.com"},
]
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["jax", "linear-algebra", "gaussian", "equinox", "lineax"]
requires-python = ">=3.12"
classifiers = [
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Natural Language :: English",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
    "Development Status :: 2 - Pre-Alpha",
]
dependencies = [
    "jax>=0.4",
    "jaxlib>=0.4",
    "equinox>=0.11",
    "lineax>=0.0.5",
    "matfree>=0.3",
    "jaxtyping>=0.2",
    "einops>=0.8",
]

[project.urls]
Repository = "https://github.com/jejjohnson/gaussx"

[project.optional-dependencies]
numpyro = [
    "numpyro>=0.15",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-xdist>=3.5",
    "pre-commit>=3.7",
    "beartype>=0.18",
    "numpyro>=0.15",
]
lint = [
    "ruff>=0.9",
]
typecheck = [
    "ty>=0.0.1a6",
    "numpyro>=0.15",
]
docs = [
    "matplotlib>=3.8",
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "mkdocs-jupyter>=0.24",
    "jupytext>=1.16",
    "ipykernel>=6.29",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gaussx"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E402", "E721", "E731", "E741", "F722"]
allowed-confusables = ["−", "σ", "𝒩", "×", "μ", "Σ", "α", "β", "λ", "ε", "∈", "⊙", "≈", "⟺"]

[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2

[tool.ty.environment]
python-version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["slow: marks tests as slow (MCMC, SVI — deselect with '-m \"not slow\"')"]

[tool.coverage.run]
source = ["src/gaussx"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
